Time: 15 ms (9.85%), Space: 8.7 MB (15.03%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    double mySqrt(double x) {
-        double ans = 1;
+    long long mySqrt(long long x) {
+        long long ans = 1;
         if(x == 1 || x == 0){
             return x;
         }
